% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{gpusvd}
\alias{gpusvd}
\alias{gpulu}
\title{cuSOLVER svd of a gpuRcudaMatrix}
\usage{
gpusvd(x, ...)

gpulu(x, PIV_FLAG = 1, ...)
}
\arguments{
\item{x}{a gpuRcudaMatrix object}
}
\description{
return the singular value decomposition of a gpuRcudaMatrix

return the LU decomposition of a gpuRcudaMatrix
}
\author{
Chaitanya Inumella.

Chaitanya Inumella.
}
