% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{\%*\%,cudaMatrix,cudaMatrix-method}
\alias{\%*\%,cudaMatrix,cudaMatrix-method}
\title{cuBLAS Matrix Multiplication}
\usage{
\S4method{\%*\%}{cudaMatrix,cudaMatrix}(x, y)
}
\arguments{
\item{x}{A gpuRcuda object}

\item{y}{A gpuRcuda object}
}
\description{
Multiply two gpuRcude objects, if they are conformable.
}
\author{
Charles Determan Jr.
}
