% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{\%*\%,cudaMatrix,cudaMatrix-method}
\alias{\%*\%,cudaMatrix,cudaMatrix-method}
\alias{svd,gpuRcudaMatrix-method}
\title{cuBLAS Matrix Multiplication}
\usage{
\S4method{\%*\%}{cudaMatrix,cudaMatrix}(x, y)

\S4method{svd}{gpuRcudaMatrix}(x)
}
\arguments{
\item{x}{A gpuRcudaMatrix object}

\item{y}{A gpuRcuda object}
}
\description{
Multiply two gpuRcude objects, if they are conformable.

return the singular value decomposition of a gpuRcudaMatrix
}
\author{
Charles Determan Jr.

Chaitanya Inumella.
}
